Move PaymentHistory component to Pages directory and remove old History directory
